Updating the "Revoke Article 50 and remain in the EU." petition data.
Data actually closer to the summary reported on the website this time around.

Total from the json file: 3,708,909 - looks like that is out of sync with the total being reported on the website, but I'm guessing that value has had to be cached to reduce load on the petitions site.

Also disabling some tests that were specific to an earlier petition.